refactor: move structs
